Изменен parser, добавлены tests2.cs, добавлены комментарии

+++ parser.cs
@@ -4,12 +4,13 @@ namespace Calculator_Parser
 {
     public class Parser
     {
-        private string InputString;
-        private bool isInvalid = false;
-        private int symbol, index = 0;
+        private string InputString;  // Входная строка для парсера
+        private bool isInvalid = false;  // Флаг для отслеживания ошибок при парсинге
+        private int symbol, index = 0;  // Переменные для отслеживания текущего символа и индекса входной строки
 
         public Parser(string? sourceString)
         {
+            // Инициализация входной строки, если она не пуста, иначе устанавливаем значение "0"
             InputString = sourceString ?? "0";
         }
 
@@ -18,7 +19,9 @@ namespace Calculator_Parser
         {
             if (index < InputString.Length)
             {
+                // Получение символа из строки
                 symbol = InputString[index];
+                // Замена символов "×" на "*" и "÷" на "/"
                 if (symbol == '×')
                 {
                     symbol = '*';
@@ -30,12 +33,12 @@ namespace Calculator_Parser
                 index++;
             }
             else
+                // Если достигнут конец строки, устанавливаем символ в '\0'
                 symbol = '\0';
         }
 
-
         // Метод для разбора входной строки и выполнения вычислений
-        public double Parse()
+        public decimal Parse()
         {
             // Проверка наличия входной строки
             if (string.IsNullOrWhiteSpace(InputString))
@@ -44,11 +47,12 @@ namespace Calculator_Parser
             // Если входная строка равна "0", вернуть 0
             if (InputString == "0")
                 return 0;
+
             // Инициализация индекса и первого символа
             index = 0;
             GetSymbol();
             // Вызов метода для разбора выражения
-            double result = MethodE();
+            decimal result = MethodE();
             // Проверка наличия ошибок в выражении
             if (isInvalid)
                 throw new InvalidOperationException("Invalid input");
@@ -56,15 +60,17 @@ namespace Calculator_Parser
             return result;
         }
 
-        private double MethodE()
+        private decimal MethodE()
         {
-            double x = MethodT();
+            // Метод для разбора выражения
+            decimal x = MethodT();
             while (symbol == '+' || symbol == '-')
             {
                 char p = (char)symbol;
                 GetSymbol();
                 if (symbol == '\0')
                 {
+                    // Если достигнут конец строки, выполнить операцию с предыдущим результатом
                     if (p == '+')
                         x += MethodT();
                     else
@@ -78,14 +84,15 @@ namespace Calculator_Parser
             return x;
         }
 
-        private double MethodT()
+        private decimal MethodT()
         {
-            double x = MethodM();
+            // Метод для разбора терма
+            decimal x = MethodM();
             while (symbol == '*' || symbol == '/')
             {
                 char p = (char)symbol;
                 GetSymbol();
-                double y = MethodM();
+                decimal y = MethodM();
 
                 if (p == '*')
                 {
@@ -95,8 +102,9 @@ namespace Calculator_Parser
                 {
                     if (y == 0)
                     {
+                        // Обработка деления на ноль
                         isInvalid = true;
-                        return 0; // Деление на ноль
+                        return 0;
                     }
                     x /= y;
                 }
@@ -104,16 +112,18 @@ namespace Calculator_Parser
             return x;
         }
 
-
-        private double MethodM()
+        private decimal MethodM()
         {
-            double x;
+            // Метод для разбора множителя
+            decimal x;
             if (symbol == '(')
             {
                 GetSymbol();
+                // Рекурсивный вызов для разбора выражения в скобках
                 x = MethodE();
                 if (symbol != ')')
                 {
+                    // Если закрывающая скобка не найдена, установить флаг ошибки
                     isInvalid = true;
                     return 0;
                 }
@@ -124,14 +134,17 @@ namespace Calculator_Parser
                 if (symbol == '-')
                 {
                     GetSymbol();
+                    // Унарный минус
                     x = -MethodM();
                 }
                 else if (symbol >= '0' && symbol <= '9')
                 {
+                    // Если символ - цифра, вызвать метод для разбора числа
                     x = MethodC();
                 }
                 else
                 {
+                    // Если символ не является цифрой, установить флаг ошибки
                     isInvalid = true;
                     return 0;
                 }
@@ -139,8 +152,9 @@ namespace Calculator_Parser
             return x;
         }
 
-        private double MethodC()
+        private decimal MethodC()
         {
+            // Метод для разбора числа
             string x = "";
             bool isDot = false;
             while (symbol >= '0' && symbol <= '9' || symbol == '.' || symbol == ',')
@@ -149,6 +163,7 @@ namespace Calculator_Parser
                 {
                     if (isDot)
                     {
+                        // Если уже встречена точка, установить флаг ошибки
                         isInvalid = true;
                         return 0;
                     }
@@ -161,9 +176,9 @@ namespace Calculator_Parser
                 }
                 GetSymbol();
             }
-            return double.Parse(x, System.Globalization.CultureInfo.InvariantCulture);
+            // Преобразование строки в десятичное число, игнорируя культуру
+            return decimal.Parse(x, System.Globalization.CultureInfo.InvariantCulture);
         }
-
     }
 }
